fix: 修复回放只能播放前655.36秒； fix: 修复VSResultsMenu的再来一局按钮在本地对战时也会3秒后才可点击

+++ app/src/main/cpp/PvZ/include/PvZ/Lawn/LawnApp.h
@@ -392,7 +392,7 @@ inline int gNetDelayNow = 0;
 inline bool gNetPingHasValidDelay = false;
 inline bool gNetPingAwaitingPong = false;
 inline int gNetPingSendCounter = 0;
-inline uint16_t gNetPingNowTick = 0;
+inline uint32_t gNetPingNowTick = 0;
 inline uint16_t gNetPingLatestSentTick = 0;
 inline uint16_t gNetPingLastPongTick = 0;
